autorize user controller

+++ moduls/Badzohreh/User/Http/Controllers/UserController.php
@@ -35,6 +35,7 @@ class UserController extends Controller
 
     public function addRole($userId, AddRoleRequest $request)
     {
+        $this->authorize("addRole",User::class);
         $user = $this->userRepo->findById($userId);
         $user->assignRole($request->role);
         showFeedbacks("عملیات موفقیت آمیز", "با موفقیت اضافه شد.");
@@ -43,6 +44,7 @@ class UserController extends Controller
 
     public function giveRole($user, $role)
     {
+        $this->authorize("removeRole",User::class);
         $user = $this->userRepo->findById($user);
         if ($user->removeRole($role)) {
             return AjaxResponses::successResponses();
@@ -53,6 +55,8 @@ class UserController extends Controller
     public function edit($userId)
     {
 
+        $this->authorize("update",User::class);
+
         $user = $this->userRepo->findById($userId);
 
         return view("User::users.edit", compact("user"));
@@ -60,6 +64,8 @@ class UserController extends Controller
 
     public function update($userId, UpdateUserRequest $request)
     {
+        $this->authorize("update",User::class);
+
         $user = $this->userRepo->findById($userId);
         if ($request->file("image")) {
             if ($user->banner) {
@@ -76,12 +82,13 @@ class UserController extends Controller
 
     public function destroy($userId)
     {
-
+        $this->authorize("delete",User::class);
         $this->userRepo->delete($userId);
     }
 
     public function manualConfirm($userId)
     {
+        $this->authorize("manaualConfirm",User::class);
         $user = $this->userRepo->findById($userId);
         $user->markEmailAsVerified();
         return AjaxResponses::successResponses();
